relative path to jest extention working
