+++ src/lib/rng/knuth-taocp/__test__/test.ts
@@ -1,74 +1,22 @@
 import { KnuthTAOCP } from '../';
 import { fixture as getRFixture } from './fixture';
+import '../../../../../packages/jest-extension';
+//import '$jest-extension';
 
 const fixture = getRFixture();
-
-expect.extend({
-    matchFloatingPointBinary(received, expected) {
-        // let received: number[]  = []
-
-        let pass = 0;
-        if (Array.isArray(received) && Array.isArray(expected) && expected.length == received.length) {
-            pass = 1;
-        } else if (typeof received === 'number' && typeof expected === 'number') {
-            pass = 2;
-        } else if (
-            received instanceof Float32Array &&
-            expected instanceof Float32Array &&
-            received.length === expected.length
-        ) {
-            pass = 3;
-        } else if (
-            received instanceof Float64Array &&
-            expected instanceof Float64Array &&
-            received.length === expected.length
-        ) {
-            pass = 4;
-        }
-
-        if (pass === 0) {
-            const options = {
-                comment: 'data types not comparable',
-                isNot: this.isNot,
-                promise: this.promise,
-            };
-            const message = () =>
-                this.utils.matcherHint('matchFloatingPointBinary', undefined, undefined, options) +
-                '\n\n' +
-                `Expected: [type] is not of equal type/length as received\n` +
-                `Received: [type] is not of equal type/length as expected`;
-            return {
-                message,
-                pass: this.isNot ? true : false,
-            };
-        }
-        return {
-            pass: true,
-            message: () => '',
-        };
-    },
-    /* 
-            Now we do the real checks.
-            }
-            const options = {
-                comment: '< inequality check',
-                isNot: this.isNot,
-                promise: this.promise,
-            };
-            const message =
-                () => this.utils.matcherHint('toBeLowerThen', undefined, undefined, options) +
-                '\n\n' +
-                `Expected: should be bigger then ${this.utils.printExpected(ceiling)}\n` +
-                `Received: ${this.utils.printReceived(received)}`
-                :
-                () => `expected ${received} to be lower then ${ceiling}`;
-          */
-});
-
+/*
+function getTypeString(o: any) {
+    const ts = typeof o;
+    if (o === undefined) return 'undefined';
+    if (o === null) return null;
+    if (['string', 'number', 'symbol', 'function'].includes(ts)) return ts;
+}
+*/
 describe('rng knuth-taocp', function n() {
     it.only('some test', () => {
         expect.hasAssertions();
-        expect(undefined).not.matchFloatingPointBinary(90);
+        const e = expect(undefined);
+        e.matchFloatingPointBinary(90);
     });
 
     it('sample for seed=0, n=10', () => {
