+++ src/lib/rng/knuth-taocp/index.ts
@@ -16,13 +16,12 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-
 import { fixup } from '../fixup';
 import { IRNG } from '../irng';
 import { IRNGTypeEnum } from '../irng-type';
 import { seed } from '../timeseed';
 import { TAOCP1997init } from './taocp-1997-init';
-import { seedCheck } from '../seedcheck'
+import { seedCheck } from '../seedcheck';
 /* helpers */
 /* helpers */
 /* helpers */
@@ -39,84 +38,83 @@ const mod_diff = (x: number, y: number) => (x - y) & (MM - 1);
 const SEED_LEN = 101;
 
 export class KnuthTAOCP extends IRNG {
-  //private buf: ArrayBuffer;
+    //private buf: ArrayBuffer;
 
-  private m_seed: Uint32Array;
+    private m_seed: Uint32Array;
 
-  private get KT_pos() {
-    return this.m_seed[100];
-  }
-
-  private set KT_pos(v: number) {
-    this.m_seed[100] = v;
-  }
-
-  private fixupSeeds(): void {
-    if (this.KT_pos <= 0) this.KT_pos = 100;
-    /* check for all zeroes */
-    const s = this.m_seed.slice(0, 100);
-    if (s.find(v => !!v) === undefined) this.init(seed());
-    return;
-  }
-
-  private KT_next() {
-    const s = this.m_seed;
-    if (this.KT_pos >= 100) {
-      this.ran_arr_cycle();
-      this.KT_pos = 0;
+    private get KT_pos() {
+        return this.m_seed[100];
     }
-    return s[this.KT_pos++];
-  }
 
-  private RNG_Init_R_KT(_seed: number) {
-    this.m_seed.set(TAOCP1997init(_seed % 1073741821));
-    this.KT_pos = 100;
-    this.fixupSeeds();
-  }
-
-  private ran_arr_cycle(): number {
-    this.ran_array(ran_arr_buf, QUALITY);
-    ran_arr_buf[KK] = -1;
-    return ran_arr_buf[0];
-  }
-
-  private ran_array(aa: Uint32Array, n: number) {
-    let i;
-    let j;
-    const ran_x = this.m_seed;
-    for (j = 0; j < KK; j++) aa[j] = ran_x[j];
-    for (; j < n; j++) aa[j] = mod_diff(aa[j - KK], aa[j - LL]);
-    for (i = 0; i < LL; i++, j++) ran_x[i] = mod_diff(aa[j - KK], aa[j - LL]);
-    for (; i < KK; i++, j++) ran_x[i] = mod_diff(aa[j - KK], ran_x[i - LL]);
-  }
-
-  public constructor(_seed = seed()) {
-    super('Knuth-TAOCP', IRNGTypeEnum.KNUTH_TAOCP);
-    this.m_seed = new Uint32Array(SEED_LEN);
-    this.init(_seed);
-  }
-
-  public init(_seed: number  =  seed()) {
-    /* Initial scrambling */
-    const s = new Uint32Array([0]);
-    s[0] = _seed;
-    for (let j = 0; j < 50; j++) {
-      s[0] = 69069 * s[0] + 1;
+    private set KT_pos(v: number) {
+        this.m_seed[100] = v;
     }
-    this.RNG_Init_R_KT(s[0]);
-  }
 
-  public set seed(_seed: Uint32Array) {
+    private fixupSeeds(): void {
+        if (this.KT_pos <= 0) this.KT_pos = 100;
+        /* check for all zeroes */
+        const s = this.m_seed.slice(0, 100);
+        if (s.find((v) => !!v) === undefined) this.init(seed());
+        return;
+    }
 
-    seedCheck(this._kind,_seed, SEED_LEN)
-    this.m_seed.set(_seed);
-  }
+    private KT_next() {
+        const s = this.m_seed;
+        if (this.KT_pos >= 100) {
+            this.ran_arr_cycle();
+            this.KT_pos = 0;
+        }
+        return s[this.KT_pos++];
+    }
 
-  internal_unif_rand(): number {
-     return fixup(this.KT_next() * KT);
-  }
+    private RNG_Init_R_KT(_seed: number) {
+        this.m_seed.set(TAOCP1997init(_seed % 1073741821));
+        this.KT_pos = 100;
+        this.fixupSeeds();
+    }
 
-  public get seed() {
-    return this.m_seed;
-  }
+    private ran_arr_cycle(): number {
+        this.ran_array(ran_arr_buf, QUALITY);
+        ran_arr_buf[KK] = -1;
+        return ran_arr_buf[0];
+    }
+
+    private ran_array(aa: Uint32Array, n: number) {
+        let i;
+        let j;
+        const ran_x = this.m_seed;
+        for (j = 0; j < KK; j++) aa[j] = ran_x[j];
+        for (; j < n; j++) aa[j] = mod_diff(aa[j - KK], aa[j - LL]);
+        for (i = 0; i < LL; i++, j++) ran_x[i] = mod_diff(aa[j - KK], aa[j - LL]);
+        for (; i < KK; i++, j++) ran_x[i] = mod_diff(aa[j - KK], ran_x[i - LL]);
+    }
+
+    public constructor(_seed = seed()) {
+        super('Knuth-TAOCP', IRNGTypeEnum.KNUTH_TAOCP);
+        this.m_seed = new Uint32Array(SEED_LEN);
+        this.init(_seed);
+    }
+
+    public init(_seed: number = seed()) {
+        /* Initial scrambling */
+        const s = new Uint32Array([0]);
+        s[0] = _seed;
+        for (let j = 0; j < 50; j++) {
+            s[0] = 69069 * s[0] + 1;
+        }
+        this.RNG_Init_R_KT(s[0]);
+    }
+
+    public set seed(_seed: Uint32Array) {
+        seedCheck(this._kind, _seed, SEED_LEN);
+        this.m_seed.set(_seed);
+    }
+
+    public get seed() {
+        return this.m_seed;
+    }
+
+    internal_unif_rand(): number {
+        return fixup(this.KT_next() * KT);
+    }
 }
