--- test/@types/jest/index.d.ts
@@ -1,7 +1,0 @@
-declare global {
-    namespace jest {
-        interface Matchers<R> {
-            matchFloatingPointBinary(expected: any): R;
-        }
-    }
-}
