--- test/@types/jest/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "jest",
-  "main":"",
-  "types":"index.d.ts"
-}
